Move image component out of slice
