+++ site/pages/what-we-do/cross-functional-slice/styles.css
@@ -49,7 +49,7 @@
   width: 260px;
 }
 
-.list {
+.colourBoxContainer {
   display: flex;
   flex-wrap: wrap;
   padding-left: 20px;
@@ -92,7 +92,7 @@ a .detail:hover {
     padding: 0;
   }
   
-  .list {
+  .colourBoxContainer {
     padding: 40px 0 0 60px;
   }
